+++ airflow/contrib/operators/sagemaker_endpoint_config_operator.py
@@ -16,51 +16,20 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""
+This module is deprecated.
+Please use `airflow.providers.amazon.aws.operators.sagemaker_endpoint_config`.
+"""
 
-from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.operators.sagemaker_base import SageMakerBaseOperator
-from airflow.utils.decorators import apply_defaults
+import warnings
 
+# pylint: disable=unused-import
+from airflow.providers.amazon.aws.operators.sagemaker_endpoint_config import (  # noqa
+    SageMakerEndpointConfigOperator,
+)
 
-class SageMakerEndpointConfigOperator(SageMakerBaseOperator):
-
-    """
-    Create a SageMaker endpoint config.
-
-    This operator returns The ARN of the endpoint config created in Amazon SageMaker
-
-    :param config: The configuration necessary to create an endpoint config.
-
-        For details of the configuration parameter see :py:meth:`SageMaker.Client.create_endpoint_config`
-    :type config: dict
-    :param aws_conn_id: The AWS connection ID to use.
-    :type aws_conn_id: str
-    """
-
-    integer_fields = [
-        ['ProductionVariants', 'InitialInstanceCount']
-    ]
-
-    @apply_defaults
-    def __init__(self,
-                 config,
-                 *args, **kwargs):
-        super().__init__(config=config,
-                         *args, **kwargs)
-
-        self.config = config
-
-    def execute(self, context):
-        self.preprocess_config()
-
-        self.log.info('Creating SageMaker Endpoint Config %s.', self.config['EndpointConfigName'])
-        response = self.hook.create_endpoint_config(self.config)
-        if response['ResponseMetadata']['HTTPStatusCode'] != 200:
-            raise AirflowException(
-                'Sagemaker endpoint config creation failed: %s' % response)
-        else:
-            return {
-                'EndpointConfig': self.hook.describe_endpoint_config(
-                    self.config['EndpointConfigName']
-                )
-            }
+warnings.warn(
+    "This module is deprecated. "
+    "Please use `airflow.providers.amazon.aws.operators.sagemaker_endpoint_config`.",
+    DeprecationWarning, stacklevel=2
+)
