+++ tests/providers/amazon/aws/operators/test_sagemaker_endpoint_config.py
@@ -22,8 +22,8 @@ import unittest
 import mock
 
 from airflow.contrib.hooks.sagemaker_hook import SageMakerHook
-from airflow.contrib.operators.sagemaker_endpoint_config_operator import SageMakerEndpointConfigOperator
 from airflow.exceptions import AirflowException
+from airflow.providers.amazon.aws.operators.sagemaker_endpoint_config import SageMakerEndpointConfigOperator
 
 model_name = 'test-model-name'
 config_name = 'test-config-name'
